refactor: 게임 진행을 담당하는 BridgeGameManager 클래스를 만들어 분리
BridgeGameManager 클래스는 Model과 View를 연결하는 Controller의 역할을 한다.
