+++ __tests__/BridgeGameTest.js
@@ -1,6 +1,6 @@
 const { describe, expect, test } = require('@jest/globals');
 const { Console } = require('@woowacourse/mission-utils');
-const App = require('../src/App');
+const BridgeGameManager = require('../src/BridgeGameManager');
 const BridgeGame = require('../src/BridgeGame');
 
 let crossingOrder = [];
@@ -59,135 +59,6 @@ describe('move 메서드 테스트', () => {
 
     expect(crossingOrder[0][1]).toEqual(0);
     crossingOrder = [];
-  });
-});
-
-describe('getBridgeCrossingResult 메서드 테스트', () => {
-  test('방향을 입력한 횟수 만큼 호출된다.', () => {
-    const app = new App();
-    const bridgeGame = new BridgeGame(['U', 'D', 'D']);
-
-    app.bridgeGame = bridgeGame;
-
-    const crossingResultspy = jest.spyOn(bridgeGame, 'getBridgeCrossingResult');
-
-    Console.readLine = jest.fn();
-    Console.readLine
-      .mockImplementationOnce((_, callback) => {
-        callback('U');
-      })
-      .mockImplementationOnce((_, callback) => {
-        callback('D');
-      });
-
-    app.requestDirection();
-
-    expect(crossingResultspy).toHaveBeenCalledTimes(2);
-  });
-
-  test('메서드가 반환한 배열은 항상 길이가 2이다.', () => {
-    const app = new App();
-    const bridgeGame = new BridgeGame(['U', 'D', 'D']);
-
-    app.bridgeGame = bridgeGame;
-
-    const crossingResultspy = jest.spyOn(bridgeGame, 'getBridgeCrossingResult');
-
-    crossingResultspy.mockClear();
-
-    Console.readLine = jest.fn();
-    Console.readLine
-      .mockImplementationOnce((_, callback) => {
-        callback('U');
-      })
-      .mockImplementationOnce((_, callback) => {
-        callback('D');
-      });
-
-    app.requestDirection();
-
-    crossingResultspy.mock.results.forEach(({ value }) => {
-      expect(value).toHaveLength(2);
-    });
-  });
-
-  test('첫번째 요소는 위쪽의 다리를 뜻하며 두번째 요소는 아래쪽의 다리를 뜻한다. 한 턴에 하나의 다리만 선택할 수 있으며 선택받지 않은 다리는 공백을 한 칸 가진 빈 문자열을 요소로 가진다.', () => {
-    const app = new App();
-    const bridgeGame = new BridgeGame(['U', 'D', 'D']);
-
-    app.bridgeGame = bridgeGame;
-
-    const crossingResultspy = jest.spyOn(bridgeGame, 'getBridgeCrossingResult');
-
-    crossingResultspy.mockClear();
-
-    Console.readLine = jest.fn();
-    Console.readLine.mockImplementationOnce((_, callback) => {
-      callback('U');
-    });
-
-    app.requestDirection();
-
-    crossingResultspy.mock.results.forEach(({ value }) => {
-      expect(value[1]).toEqual([' ']);
-    });
-  });
-
-  test('건널 수 있는 다리는 O를 표시한다.', () => {
-    const app = new App();
-    const bridgeGame = new BridgeGame(['U', 'D', 'D']);
-
-    app.bridgeGame = bridgeGame;
-
-    const crossingResultspy = jest.spyOn(bridgeGame, 'getBridgeCrossingResult');
-
-    crossingResultspy.mockClear();
-
-    Console.readLine = jest.fn();
-    Console.readLine
-      .mockImplementationOnce((_, callback) => {
-        callback('U');
-      })
-      .mockImplementationOnce((_, callback) => {
-        callback('D');
-      });
-
-    app.requestDirection();
-
-    const lastIdx = crossingResultspy.mock.results.length - 1;
-
-    const result = crossingResultspy.mock.results[lastIdx].value;
-
-    expect(result[0][0]).toEqual('O');
-    expect(result[1][1]).toEqual('O');
-  });
-
-  test('건널 수 없는 다리는 X를 표시한다.', () => {
-    const app = new App();
-    const bridgeGame = new BridgeGame(['U', 'D', 'D']);
-
-    app.bridgeGame = bridgeGame;
-
-    const crossingResultspy = jest.spyOn(bridgeGame, 'getBridgeCrossingResult');
-
-    crossingResultspy.mockClear();
-
-    Console.readLine = jest.fn();
-    Console.readLine
-      .mockImplementationOnce((_, callback) => {
-        callback('U');
-      })
-      .mockImplementationOnce((_, callback) => {
-        callback('U');
-      });
-
-    app.requestDirection();
-
-    const lastIdx = crossingResultspy.mock.results.length - 1;
-
-    const result = crossingResultspy.mock.results[lastIdx].value;
-
-    expect(result[0][1]).toEqual('X');
   });
 });
 
@@ -262,10 +133,10 @@ describe('getResult 메서드 테스트', () => {
     [['U', 'U'], false],
     [['U', 'D', 'D', 'U'], true],
   ])('재시도가 없을 때의 반환값 테스트', (array, boolean) => {
-    const app = new App();
+    const bridgeGameManager = new BridgeGameManager();
     const bridgeGame = new BridgeGame(['U', 'D', 'D', 'U']);
 
-    app.bridgeGame = bridgeGame;
+    bridgeGameManager.bridgeGame = bridgeGame;
 
     crossingOrder = [];
 
@@ -282,10 +153,10 @@ describe('getResult 메서드 테스트', () => {
   });
 
   test('재시도가 있을 때의 반환값 테스트', () => {
-    const app = new App();
+    const bridgeGameManager = new BridgeGameManager();
     const bridgeGame = new BridgeGame(['U', 'D', 'D', 'U']);
 
-    app.bridgeGame = bridgeGame;
+    bridgeGameManager.bridgeGame = bridgeGame;
 
     crossingOrder = [];
