+++ src/App.js
@@ -1,98 +1,14 @@
-const { Console } = require('@woowacourse/mission-utils');
-const BridgeGame = require('./BridgeGame');
-const BridgeMaker = require('./BridgeMaker');
-const { generate } = require('./BridgeRandomNumberGenerator');
-const InputView = require('./View/InputView');
-const OutputView = require('./View/OutputView');
-const Validation = require('./Validation');
-const { throwException } = require('./ErrorHandler');
+const BridgeGameManager = require('./BridgeGameManager');
 
 class App {
+  #bridgeGameManager;
+
   constructor() {
-    this.bridgeGame = null;
+    this.#bridgeGameManager = new BridgeGameManager();
   }
 
   play() {
-    OutputView.printStartMessage();
-
-    this.requestBridgeSize();
-  }
-
-  requestBridgeSize() {
-    InputView.readBridgeSize((size) => {
-      const { errorMsg } = Validation.checkBridgeSize(size);
-      if (errorMsg)
-        return throwException(errorMsg, () => this.requestBridgeSize());
-
-      this.createBridgeGame(size);
-
-      this.requestDirection();
-    });
-  }
-
-  createBridgeGame(size) {
-    const bridge = BridgeMaker.makeBridge(Number(size), generate);
-
-    this.bridgeGame = new BridgeGame(bridge);
-  }
-
-  requestDirection() {
-    InputView.readMoving((direction) => {
-      const { errorMsg } = Validation.checkDirection(direction);
-      if (errorMsg)
-        return throwException(errorMsg, () => this.requestDirection());
-
-      this.bridgeGame.move(direction);
-      this.printBridgeCrossingResult();
-
-      this.actionAboutBridgeGame();
-    });
-  }
-
-  printBridgeCrossingResult() {
-    const bridgeCrossingResult = this.bridgeGame.getBridgeCrossingResult();
-
-    OutputView.printMap(bridgeCrossingResult);
-  }
-
-  actionAboutBridgeGame() {
-    if (this.bridgeGame.isFail()) return this.requestRestartOrQuit();
-
-    if (this.bridgeGame.isLast()) return this.quit();
-
-    return this.requestDirection();
-  }
-
-  requestRestartOrQuit() {
-    InputView.readGameCommand((commandOption) => {
-      const { errorMsg } = Validation.checkCommandOption(commandOption);
-      if (errorMsg)
-        return throwException(errorMsg, () => this.requestRestartOrQuit());
-
-      this.actionAboutGameCommand(commandOption);
-    });
-  }
-
-  actionAboutGameCommand(commandOption) {
-    if (commandOption === 'R') return this.restart();
-
-    return this.quit();
-  }
-
-  restart() {
-    this.bridgeGame.retry();
-    this.requestDirection();
-  }
-
-  quit() {
-    this.printGameResult();
-    Console.close();
-  }
-
-  printGameResult() {
-    OutputView.printEndMessage(this.bridgeGame.isFail());
-    OutputView.printMap(this.bridgeGame.getBridgeCrossingResult());
-    OutputView.printResult(this.bridgeGame.getResult());
+    this.#bridgeGameManager.start();
   }
 }
